scripts/requirements.txt
markitdown>=0.1.5
anthropic>=0.40.0
openai>=1.0.0
pyyaml>=6.0